+++ spreadsheets/.clasp.json
@@ -1,1 +1,4 @@
-{"scriptId":"1-l-C3Axp4PRpogfJNjCk1m7qss9TSymsPujhRmbjxnGoNn4ISYxDuryb"}
+{
+  "scriptId": "1-l-C3Axp4PRpogfJNjCk1m7qss9TSymsPujhRmbjxnGoNn4ISYxDuryb",
+  "rootDir": "dist"
+}
