+++ spreadsheets/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "spreadsheets",
+  "version": "1.0.0",
+  "scripts": {
+    "dev": "babel src --watch -d dist",
+    "build": "babel src --compact=true -d dist --no-comments"
+  },
+  "main": "index.js",
+  "license": "MIT",
+  "devDependencies": {
+    "@babel/cli": "^7.4.3",
+    "@babel/core": "^7.4.3",
+    "@babel/plugin-proposal-object-rest-spread": "^7.4.3",
+    "@babel/preset-env": "^7.4.3",
+    "babel-plugin-dotenv": "^0.1.1",
+    "eslint": "^5.16.0",
+    "eslint-config-prettier": "^4.1.0",
+    "eslint-plugin-googleappsscript": "^1.0.1",
+    "eslint-plugin-prettier": "^3.0.1",
+    "prettier": "^1.17.0"
+  }
+}
